Introduce DSL traits with deprecated forwarders
Extract the duplicated DSL surface across DefinitionBuilder,
PropertyTypeDeclaration, and BuildModelDeclaration into a small set
of Groovy traits under builders/dsl/: HasProperties, HasNestedTypes,
HasInjectedServices, HasAnnotations, HasUndiscoverableNested, and two
shared-ref traits whose storage location differs by container.
