+++ platforms/core-configuration/declarative-dsl-provider/src/testFixtures/groovy/org/gradle/features/internal/builders/dsl/HasSharedRefInNestedTypes.groovy
@@ -1,0 +1,44 @@
+/*
+ * Copyright 2026 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.gradle.features.internal.builders.dsl
+
+import org.gradle.features.internal.builders.PropertyTypeDeclaration
+
+/**
+ * Contributes {@code sharedProperty(...)} that stores a shared-type reference as a
+ * {@link PropertyTypeDeclaration} entry in {@link #getNestedTypes()}.
+ *
+ * <p>Use on contexts where the shared-typed property is rendered as a {@code @Nested}
+ * getter returning the shared type's class directly (e.g. {@code DefinitionBuilder}).</p>
+ */
+trait HasSharedRefInNestedTypes {
+    abstract List<PropertyTypeDeclaration> getNestedTypes()
+
+    /** Adds a property whose type is a previously declared shared type. */
+    void sharedProperty(String name, PropertyTypeDeclaration ref) {
+        nestedTypes.add(new PropertyTypeDeclaration(
+            name: name,
+            typeName: ref.typeName,
+            properties: ref.properties,
+            nestedTypes: ref.nestedTypes,
+            injectedServices: ref.injectedServices,
+            implementsDefinition: ref.implementsDefinition,
+            buildModel: ref.buildModel,
+            isSharedRef: true
+        ))
+    }
+}
